Asgn-2 Done

+++ 2/include/case1.hpp
@@ -1,4 +1,4 @@
 using namespace std;
 
-void storeInput(unordered_map<long long int, nodes> &nodesList, vector<ways> &waysList);
-void searchInput(unordered_map<long long int, nodes> &nodesList, string input);
+void storeInput(unordered_map<ll, nodes> &nodesList, vector<ways> &waysList); // will store list of nodes in nodesList and list of ways in waysList
+void searchInput(unordered_map<ll, nodes> &nodesList, string input); // search using matching substrings
