+++ 2/include/ways.hpp
@@ -1,10 +1,11 @@
 using namespace std;
 
+// a class to store various attributes of ways
 class ways
 {
 public:
     string name;
-    long long int id;
-    vector<nodes> wayPoints;
+    ll id;
+    vector<nodes> wayPoints; // to store list of ways on the node
 };
